Create decks/name.rs
